requirements.txt
Flask>=3.0
Flask-Session>=0.6
msal>=1.28
requests>=2.31
python-dotenv>=1.0
